# Remove extra allocations from perf.Reader.Read()