--- a/internal/endian.go
+++ b/internal/endian.go
@@ -1,3 +1,6 @@
+//go:build !arm64 && !amd64
+// +build !arm64,!amd64
+
 package internal
 
 import (