--- a/internal/endian_le.go
+++ b/internal/endian_le.go
@@ -1,0 +1,10 @@
+//go:build amd64 || arm64
+// +build amd64 arm64
+
+package internal
+
+import "encoding/binary"
+
+var NativeEndian binary.ByteOrder = binary.LittleEndian
+
+const ClangEndian = "el"