Before:
```
goos: linux
goarch: arm64
pkg: github.com/cilium/ebpf/perf
BenchmarkReader
BenchmarkReader-4   	  287356	      3732 ns/op	     409 B/op	       6 allocs/op
BenchmarkReader-4   	  305671	      3701 ns/op	     409 B/op	       6 allocs/op
BenchmarkReader-4   	  305148	      3669 ns/op	     409 B/op	       6 allocs/op
BenchmarkReader-4   	  306469	      3675 ns/op	     409 B/op	       6 allocs/op
BenchmarkReader-4   	  295496	      3691 ns/op	     409 B/op	       6 allocs/op
BenchmarkReader-4   	  297704	      3710 ns/op	     410 B/op	       6 allocs/op
BenchmarkReader-4   	  302049	      3667 ns/op	     410 B/op	       6 allocs/op
BenchmarkReader-4   	  299085	      3680 ns/op	     409 B/op	       6 allocs/op
BenchmarkReader-4   	  305419	      3673 ns/op	     409 B/op	       6 allocs/op
BenchmarkReader-4   	  305707	      3691 ns/op	     409 B/op	       6 allocs/op
PASS
ok  	github.com/cilium/ebpf/perf	11.532s
```

After:
```
goos: linux
goarch: arm64
pkg: github.com/cilium/ebpf/perf
BenchmarkReader
BenchmarkReader-4   	  318324	      3476 ns/op	     384 B/op	       2 allocs/op
BenchmarkReader-4   	  330291	      3492 ns/op	     384 B/op	       2 allocs/op
BenchmarkReader-4   	  325369	      3509 ns/op	     384 B/op	       2 allocs/op
BenchmarkReader-4   	  325884	      3508 ns/op	     384 B/op	       2 allocs/op
BenchmarkReader-4   	  327452	      3497 ns/op	     384 B/op	       2 allocs/op
BenchmarkReader-4   	  326408	      3599 ns/op	     384 B/op	       2 allocs/op
BenchmarkReader-4   	  320997	      3507 ns/op	     384 B/op	       2 allocs/op
BenchmarkReader-4   	  320274	      3512 ns/op	     384 B/op	       2 allocs/op
BenchmarkReader-4   	  318555	      3506 ns/op	     384 B/op	       2 allocs/op
BenchmarkReader-4   	  329978	      3499 ns/op	     384 B/op	       2 allocs/op
PASS
ok  	github.com/cilium/ebpf/perf	11.788s
```

The remaining 2 allocs are:
1. The creation of the raw sample slice
2. The output slice from `program.Test`